Some fixes for x86_64 Some compiler warnings removed.
git-svn-id: https://svn.aros.org/svn/aros/trunk@33421 fb15a70f-31f2-0310-bbcc-cdcc74a49acc

+++ workbench/c/GfxControl.c
@@ -99,11 +99,11 @@ AROS_UFH3(LONG, PatchTask,
     if (port) {
         port->mp_Node.ln_Name = PORT_NAME;
         AddPort(port);
-        orig_func = SetFunction(CyberGfxBase, -28*LIB_VECTSIZE, MyLockBitMapTagList);
+        orig_func = SetFunction(CyberGfxBase, -28 * (WORD)LIB_VECTSIZE, MyLockBitMapTagList);
 	
 	/* Just wait for a signal from the port. There's no need to look at message contents */
 	WaitPort(port);
-	SetFunction(CyberGfxBase, -28*LIB_VECTSIZE, orig_func);
+	SetFunction(CyberGfxBase, -28 * (WORD)LIB_VECTSIZE, orig_func);
 	RemPort(port);
 	DeleteMsgPort(port);
     }
